Show areas on map
For the London Areas that have a polygon(s), we pass that through to the containing div as stimulus data, and then when that div is clicked, remove any existing layer, and add that geojson as a layer.

Zooms the map to the right level to fit the counds of that new layer.